Python/cloud_functions/Compute_Func/src/requirements.txt
google-cloud-secret-manager==2.12.6
google-cloud-bigquery==3.5.0
python-dotenv==0.21.0